Working classes, board display, movable piece_spec
At this stage, the basic building blocks of the chess game are now
functional. The player,piece, and board classes are used by main.rb to
deploy the main game loop. The game_loop establishes an instance of the
Board class, which itself consists of two Player instances, as well as
Piece instances for each of the 31 game pieces. The most important part
of the program is the legal_moves method in the Board class. This method
takes a Piece object, analyzes the status of board squares available to
the Piece based on its type(how a chess piece moves), and returns a list
of 2 element arrays representing the board squares that are currently
available to the Piece. This list is used to highlight the available
moves of a selected piece, and limits the ability of the user to only
select a target location that is contained within the list.

Currently, the game launches in a default state, alternates between user
turns, and allows for the movement of game pieces to board squares that
are contained in the Piece's available_moves. Capturing of opponent pieces
is allowed. These captured pieces are added to a list of captured pieces,
and have their board position set to nil when captured.

There is currently no win condition, as methods for check and checkmate
are still to be added. Getting the legal_moves method to be functional
for all Piece types was quite a challenge, and my original approach
introduced severe bugs for the pieces whose legal_moves code required
loops(rook, bishop, queen). These bugs were fixed by specifically referring
to [move[0], move[1]] as opposed to simply appending move to the available_moves
array. I have left the buggy code in this commit, commented out, for
posterity's sake. Next steps will be cleaning up this unused code, and
writing methods and tests for determining check and checkmate.